added basic authentication to ultical.
The authentication is handlede through http basic authentication. A
respective header is required to access protected resources. A resource
is protected if it takes the parameters `@Auth @NotNull User user`. This
sequence injects the User object that corresponds to the provided
http-request header and throws an UnAuthorized Exception if this object
is null.
I just checked that the code compiles. I did not test it!