Made the ProfileUpdate Job work.
In order for the ProfileUpdates to work it was necessary to do the following. First, the @OnApplicationStart annotation had to be removed from the update-job, as application start, in terms of the scheduler, is ways before the application's configuration has been parsed and is availble. Second, the ServiceLocator for performing the di-magic had to be made available for use in non-jersey resources. This has been achieved by the ServiceLocatorProvider singleton. Third, DfvPorfileLoader had to be implemented. I did this by creating a TempInitResource instance and injecting the required dependencies through the service-locator. Last, the serviceLocatorProvider had to be configured properly and the ProfileLoader had to be triggered, to load the initial profiles. Therefore, I registered a dedicated Feature with jersey.