Added a basic MailClient implementation to ultical.
The MailClient could be used to send UlticalMessages using an smtp-account,
that has to be provided in the UlticalConfig. The MailClient's functionality
could be tested using a simple unit-test using the GreenMail mail-server.
However, to ensure that the configuration works in real life, a health-check
has been added that sends a 'I'm alive'-message to the team.