added a first fix for issue #2
The DivisionAge has been enriched with a ageDifference attribute and the information, whether or not the players have to be older or younger than the specified difference. In the RosterResource it will be checked that only these players could be added to a Roster which are not too old or yet too young.

Furhter, the now somehow complex logic has been verified using Unit-Tests that caputre at lease the eligibility-rules for a roster. Test-cases for players, not yet known to the dataStore have not been added in order to keep the test fixture small. Eventually the method has to be refactored in order to make the client-stuff more testable.