extracted registration handling out of registerresource into handler.
Introduced a new RegistrationHandler which offers a single method registerUser(RegisterRequest). The handler
either returns the registered User instance or throws a RegistrationException in case something went wrong. The
register Response got much simpler as it only calls the RegistrationHandler and deals with it's outcome to
generate a proper RegisterResponse.
In the DfvBundle all DFV specific parts are bundled together. The bundle adds them to the environment, iff
the UlticalConfig says to do so. Otherwise this bundle does nothing. To make this work it has been necessary to
discard direct registratrion of the JobBundle, but do the JobBundles 'magic' within the DfvBundle. Fortunately,
this could be achieved within three lines of code, thus we do not have a massive code duplication here.
Finally, the DfvBinder, which is registered to Jersey by the DfvBundle, will bind the DfvRegistrationHandler to
the Registrationhandler contract with a rank of five, such that it could be injected into the RegisterResource.
What is missing is a Non-dfv registrationhandler instance and its registration in the Application with a rank
of one. The low rank is required as to ensure that the DfvRegistrationHandler will be used, if it is available.
Further, the Search-resource and the RosterResource still have to be freed from the Dfv-specific stuff...