liquibase script for discussed new class structure
introduced an untested draft of the new database schema. Mainly, a new
table REGISTERED_PLAYER is introduced and the existing tables DFV_PLAYER and
UNREGISTERED_PLAYER were removed. However, this removal is done in a way, tha
avoids loss of data and could be applied to a running, productive system.
Further, the column dfv_number is only available, if the user sets the context
"dfv" explicitly. If not, the column does not exist.

For the table DFV_MV_NAMES a good solution still has to be found. Eventually
I will a dropTable changeSet for the context !dfv (i.e. the table will be
dropped, if we are not in the dfv-context). At the moment this is only
possiblity to bring the non-dfv features live, and making further use of
liquibase.