The DfvProfileLoader now takes care of the new lastModified date in DfvMvName instances.
If the lastModified Date for a DfvMvName is newer than the lastModified date of the corresponding DfvPlayer, the DfvProfileLoader now updates the DfvMvPlayer.
The implementation is already completely functional but there are still some minor (or perhaps major) performance improvements possible as the computation of the DfvPlayers instances to update, is done within the java-code currently but I assume that the db would be much faster. Further the WebTarget used to update the DfvPlayer is recreated for each player. A template mechanism would help a lot here.