added logging configuration for issue #53
logging is done in two files: logs/ultical.log and logs/ultical-err.log. Both files are rolled over each day and zipped. The log archives will be kept for seven respective 14 days.
The format of the log-output is basically the same as the dw-default but the MDC with key remoteIp has been added to always know, who sent the request that caused the log-entry.