Ultical-admins get mailed, if player could not be removed from roster.
In case the blocking-dates contains one date, that lies in the past, the
affected player will not be removed from the roster. But the ultical-team
will be informed, that this rare case really happened.

Improved the DfvProfileLoaderTestCase such that it also checks the new methods
and added useful verify-statements in order to check that the right methods are
invoked using the correct arguments. Thanks mockito.